# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garmin Analyzer API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/activities": {
      "get": {
        "tags": [
          "activities"
        ],
        "summary": "List Activities",
        "description": "Activities newest first, optionally only one sport.",
        "operationId": "listActivities",
        "parameters": [
          {
            "name": "sport",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "run",
                    "bike",
                    "other"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sport"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/activities/{activity_id}": {
      "get": {
        "tags": [
          "activities"
        ],
        "summary": "Get Activity",
        "operationId": "getActivity",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivitySummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sync/status": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Status",
        "operationId": "getSyncStatus",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          }
        }
      }
    },
    "/sync": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Run Sync",
        "description": "Fetch new activities from Garmin now.",
        "operationId": "runSync",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultOut"
                }
              }
            }
          },
          "502": {
            "description": "Garmin login or fetch failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarminProblem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivityPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivitySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ActivityPage"
      },
      "ActivitySummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "garmin_id": {
            "type": "integer",
            "title": "Garmin Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sport": {
            "type": "string",
            "title": "Sport"
          },
          "activity_type": {
            "type": "string",
            "title": "Activity Type"
          },
          "start_time_local": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time Local"
          },
          "distance_mi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Mi"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "moving_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moving Time S"
          },
          "avg_pace_s_per_mi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Pace S Per Mi"
          },
          "elevation_gain_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation Gain Ft"
          },
          "avg_hr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Hr"
          },
          "max_hr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Hr"
          },
          "avg_cadence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cadence"
          },
          "avg_power": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Power"
          },
          "avg_stride_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Stride M"
          },
          "aerobic_te": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aerobic Te"
          },
          "anaerobic_te": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anaerobic Te"
          },
          "calories": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories"
          }
        },
        "type": "object",
        "required": [
          "id",
          "garmin_id",
          "name",
          "sport",
          "activity_type",
          "start_time_local",
          "distance_mi",
          "duration_s",
          "moving_time_s",
          "avg_pace_s_per_mi",
          "elevation_gain_ft",
          "avg_hr",
          "max_hr",
          "avg_cadence",
          "avg_power",
          "avg_stride_m",
          "aerobic_te",
          "anaerobic_te",
          "calories"
        ],
        "title": "ActivitySummary"
      },
      "GarminProblem": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "GarminProblem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          },
          "database": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ],
            "title": "Database"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "database",
          "version"
        ],
        "title": "Health"
      },
      "SyncResultOut": {
        "properties": {
          "added": {
            "type": "integer",
            "title": "Added"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "added",
          "updated",
          "total"
        ],
        "title": "SyncResultOut"
      },
      "SyncStatus": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "activity_count": {
            "type": "integer",
            "title": "Activity Count"
          }
        },
        "type": "object",
        "required": [
          "configured",
          "display_name",
          "last_sync_at",
          "activity_count"
        ],
        "title": "SyncStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}